% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtaux-package.R
\docType{package}
\name{mtaux-package}
\alias{mtaux}
\alias{mtaux-package}
\title{mtaux: Master thesis auxiliary package}
\description{
Just an auxiliary package with utils, fonts and themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/netique/mtaux}
  \item Report bugs at \url{https://github.com/netique/mtaux/issues}
}

}
\author{
\strong{Maintainer}: Jan Netik \email{netikja@gmail.com}

}
\keyword{internal}
